Reogranisation based on topics to include a random subpackage and moving relevant utility functions from the utilities subpackage which is geared more towards internal utilities and package infrastructure
