+++ quantecon/random/__init__.py
@@ -1,0 +1,11 @@
+"""
+Random SubPackage
+=================
+
+Utilities
+---------
+	Utilities to Support Generation of Random Arrays or Matrices
+
+"""
+
+from .utilities import probvec, sample_without_replacement
