+++ quantecon/util/__init__.py
@@ -4,5 +4,3 @@ API for QuantEcon Utilities
 
 from .array import searchsorted
 from .random import check_random_state
-from .random import probvec as random_probvec
-from .random import sample_without_replacement as random_sample_without_replacement
